pyproject.toml
[project]
name = "rag-mcp"
version = "0.1.0"
description = "Universal RAG MCP Server — index markdown knowledge bases for search via MCP tools and a web UI"
requires-python = ">=3.10,<3.14"
dependencies = [
    "mcp[cli]>=1.0.0",
    "lancedb>=0.20.0",
    "fastembed>=0.3.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "starlette>=0.40.0",
    "jinja2>=3.1.0",
    "uvicorn>=0.30.0",
    "markdown>=3.5",
    "pylance>=2.0.1",
]

[project.scripts]
rag-mcp = "rag_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_mcp"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
